+++ test/integration/CustomerTest.php
@@ -338,8 +338,6 @@ class CustomerTest extends BasePaymentTest
      *
      * @throws HeidelpayApiException
      * @throws \RuntimeException
-     *
-     * @group skip
      */
     public function customerShouldBeFetchedByCustomerIdAndUpdatedIfItAlreadyExists()
     {